+++ conda-recipe/bld.bat
@@ -1,0 +1,2 @@
+"%PYTHON%" install.py
+if errorlevel 1 exit 1
