+++ conda-recipe/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+$PYTHON install.py
+cp configure-nbextensions.py $PREFIX/share/jupyter/configure-nbextensions.py
