+++ conda-recipe/post-link.bat
@@ -1,0 +1,2 @@
+"%PYTHON%" $PREFIX/share/jupyter/configure-nbextensions.py debug
+if errorlevel 1 exit 1
