+++ conda-recipe/post-link.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+$PYTHON $PREFIX/share/jupyter/configure-nbextensions.py debug
+
