--- jupyter_nbconvert_config.py
@@ -1,8 +1,0 @@
-from jupyter_core.paths import jupyter_config_dir, jupyter_data_dir
-import os
-import sys
-
-sys.path.append(os.path.join(jupyter_data_dir(), 'extensions'))
-
-c = get_config()
-c.Exporter.template_path = [ '.', os.path.join(jupyter_data_dir(), 'templates') ]
