--- meta.yaml
@@ -1,39 +1,0 @@
-package:
-  name: nbextensions
-  version: "master"
-
-source:
-  git_url: https://github.com/ipython-contrib/IPython-notebook-extensions
-  git_tag: master
-
-build:
-  script: python setup.py install
-  
-requirements:
-  build:
-    - ipython >=4
-    - jupyter_client
-    - jupyter_core
-    - notebook
-    - nbconvert
-    - nbformat
-    - python
-    - traitlets
-    - psutil
-
-  run:
-    - ipython >=4
-    - jupyter_client
-    - jupyter_core
-    - notebook
-    - nbconvert
-    - nbformat
-    - python
-    - traitlets
-    - pyyaml
-
-about:
-  home: https://github.com/ipython-contrib/IPython-notebook-extensions
-  license:  Modified BSD License
-  summary: 'Notebook extensions for the Jupyter/IPython notebook'
-
